requirements.txt
numpy
matplotlib
scikit-learn
Pillow
joblib
